add relative path to component links in static builds
In a static build the links to components now begin with `./` incase
a user has set a handle with a `:` in it.

Having a `:` in your component handle would produce a static url like:
`href="foo:btn.html"`
The browser tries to interpret this as an app link, which gives you a
`Failed to launch 'foo:btn' because the scheme does not have a registered handler`
error message in the browser developer console and stops page navigation.

Having a relative path current folder `./`, so producing
`href="./foo:btn.html"` stops the browser interpreting as an app link.

For server builds, this doesn't apply, the existing paths are used as before.

Why use `:` in a component handle?

Components may naturally have `:` in their name, such as `ratio-box--16:9` 
or a user maybe name spacing components in a multisite Fractal library, 
where components may have the same name but live in different sites:

```
{% include '@foo:btn' with data %} // outputs `btn` from site `foo`, "handle": "foo:btn"
{% include '@bar:btn' with data %} // outputs `btn` from site `bar`, "handle": "bar:btn"
```