removed chalk dependency
This removes the chalk dependency by making the decision to
print to the console without color. It would have been possible
to remain on chalk 4.X.X for now, but 5.0.1 is the newest version
and has breaking changes (requiring ESM) and colors are not
something which are absolutely critical to the application.