removed `log-update` dependency
this has the effect that when running the `web.build.js` the command
line will no longer update the `⚑ Exported 645 of 807 items` line on
the console in place, but will instead print out one line on the log
for each exported item.

This means that the console output is much longer for that command,
but removing a dependency (which cannot be updated to the latest major
version because of an ESM requirement) seems like a valid tradeoff
in this instance.