removed get-ports dependency and automatic port discovery feature
This removes the `get-ports` dependency because the newest major
versions require ESM. In order to remove this dependency, the port
autodetection feature was also removed. The default port is still
3000, but if 3000 is blocked, the process logs the error and it is
the developer's responsibility to specify a new port with the
`--port` parameter.

In my experience, I am so used to using a fractal instance at a
given port that if the fractal managed to start and was running under
a different port, I probably wouldn't notice it. Explicitly logging
the error would make it easier for me to find the fractal instance.